--- README.md
@@ -1,2 +1,0 @@
-# Sudoku_V1
-Ce projet a pour but de résoudre une grille de sudoku.
